requirements.txt
Pillow
python-dotenv
discord.py
Pillow
fastapi
uvicorn

